Reduce testing symbols in production binaries (#55360)
* Move lib/srv/mock to lib/srv/mock_test.go

* remove testify from lib/srv/db/redis

* remove lib/auth/test package

* Move lib/utils/testhelpers.go to lib/utils/testutils

* duplicate auto upgrades server mock

* remove testing from fake podman server

* remove lib/fixtures/plugin.go

* remove test assertions from lib/fixtures

* remove depguard exceptions that no longer apply

* fix: comments
