+++ lib/tbot/workloadidentity/workloadattest/podman/client_test.go
@@ -29,7 +29,8 @@ import (
 )
 
 func TestClient(t *testing.T) {
-	server := podman.NewTestServer(t,
+	server, err := podman.NewFakeServer(
+		t.TempDir(),
 		podman.WithContainer(podman.Container{
 			ID:   "container-1234",
 			Name: "web-server",
@@ -45,6 +46,14 @@ func TestClient(t *testing.T) {
 			Labels: map[string]string{"technology": "node.js"},
 		}),
 	)
+	require.NoError(t, err)
+
+	server.Start()
+	t.Cleanup(func() {
+		if err := server.Close(); err != nil {
+			t.Logf("failed to close http server: %v", err)
+		}
+	})
 
 	client, err := podman.NewClient(server.Addr())
 	require.NoError(t, err)
